+++ src/sql/workbench/contrib/notebook/test/browser/notebookEditor.test.ts
@@ -6,8 +6,10 @@ import * as assert from 'assert';
 import { Deferred as DeferredPromise } from 'sql/base/common/promise';
 import { QueryTextEditor } from 'sql/workbench/browser/modelComponents/queryTextEditor';
 import { UntitledNotebookInput } from 'sql/workbench/contrib/notebook/browser/models/untitledNotebookInput';
+import { NotebookEditor } from 'sql/workbench/contrib/notebook/browser/notebookEditor';
 import { NBTestQueryManagementService } from 'sql/workbench/contrib/notebook/test/nbTestQueryManagementService';
-import { CellEditorProviderStub, NotebookEditorStub, NotebookModelStub } from 'sql/workbench/contrib/notebook/test/stubs';
+import { NotebookModelStub } from 'sql/workbench/contrib/notebook/test/stubs';
+import { TestNotebookEditor } from 'sql/workbench/contrib/notebook/test/testCommon';
 import { ICellModel } from 'sql/workbench/services/notebook/browser/models/modelInterfaces';
 import { INotebookService, NotebookRange } from 'sql/workbench/services/notebook/browser/notebookService';
 import { NotebookService } from 'sql/workbench/services/notebook/browser/notebookServiceImpl';
@@ -43,29 +45,7 @@ import { IWorkbenchThemeService } from 'vs/workbench/services/themes/common/work
 import { UntitledTextEditorInput } from 'vs/workbench/services/untitled/common/untitledTextEditorInput';
 import { IUntitledTextEditorService } from 'vs/workbench/services/untitled/common/untitledTextEditorService';
 import { workbenchInstantiationService } from 'vs/workbench/test/browser/workbenchTestServices';
-import { NotebookEditor } from 'sql/workbench/contrib/notebook/browser/notebookEditor';
 
-class TestNotebookEditor extends NotebookEditorStub {
-	constructor(private _cellGuid?: string, private _editor?: QueryTextEditor) {
-		super();
-	}
-	cellEditors: CellEditorProviderStub[] = [new TestCellEditorProvider(this._cellGuid, this._editor)];
-}
-
-class TestCellEditorProvider extends CellEditorProviderStub {
-	constructor(private _cellGuid: string, private _editor: QueryTextEditor) {
-		super();
-		let div = dom.$('div', undefined, dom.$('span', { id: 'demospan' }));
-		let firstChild = div.firstChild as HTMLElement;
-		this._editor.create(firstChild);
-	}
-	cellGuid(): string {
-		return this._cellGuid;
-	}
-	getEditor(): QueryTextEditor {
-		return this._editor;
-	}
-}
 
 suite('Test class NotebookEditor', () => {
 
@@ -124,7 +104,7 @@ suite('Test class NotebookEditor', () => {
 		instantiationService.get(IEditorService),
 		instantiationService.get(IConfigurationService)
 	);
-	const testNotebookEditor = new TestNotebookEditor(cellTextEditorGuid, queryTextEditor);
+	const testNotebookEditor = new TestNotebookEditor({ cellGuid: cellTextEditorGuid, editor: queryTextEditor });
 	testNotebookEditor.id = untitledNotebookInput.notebookUri.toString();
 	testNotebookEditor.model = new NotebookModelStub();
 	notebookService.addNotebookEditor(testNotebookEditor);
@@ -192,10 +172,10 @@ suite('Test class NotebookEditor', () => {
 		});
 	});
 
-	test('NotebookEditor-getCellEditor: Positive Tests getCellEditor() returns text editor object for valid guild input', async () => {
+	test('NotebookEditor-getCellEditor: Positive Tests getCellEditor() returns text editor object for valid guid input', async () => {
 		await setupPromise;
 		const result = notebookEditor.getCellEditor(cellTextEditorGuid);
-		assert.strictEqual(result, queryTextEditor, `notebookEditor.getCellEditor() should return ${queryTextEditor} when ${cellTextEditorGuid} is passed in for a notebookEditor of an empty document.`);
+		assert.strictEqual(result, queryTextEditor, 'notebookEditor.getCellEditor() should return an expected QueryTextEditor when a guid corresponding to that editor is passed in.');
 
 	});
